requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
ariadne==0.20.1
graphql-core==3.2.3
google-generativeai==0.3.2
reportlab==4.0.7
aiofiles==23.2.1
requests==2.31.0
httpx==0.25.2
jinja2==3.1.2
pydantic==2.5.0
python-multipart==0.0.6
starlette==0.27.0
